Fix version parsing for the following format: major.minor.bugfix.hotfix-buildnumber.
The format mentioned above with a dash separating the build number from the version would not parse correctly since the build number is parsed as a number and thus not seen as a qualifier. This would lead to the version being split into major, minor, bugfix, hotfix-buildnumber which failed to convert to a number.
Now, instead of using the original when no qualifier is found, the version string with the dash replaced with a dot is used.
Extended tests.